+++ Package.swift
@@ -11,7 +11,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/ReachFive/reachfive-ios.git", "9.0.0"..<"100.0.0"),
-        .package(url: "https://github.com/facebook/facebook-ios-sdk.git", .upToNextMinor(from: "17.4.0")),
+        .package(url: "https://github.com/facebook/facebook-ios-sdk.git", .upToNextMinor(from: "18.1.0")),
     ],
     targets: [
         .target(